+++ e2e/setup-files/fetched-data.js
@@ -5,13 +5,6 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-function* generatorMethod() {
-  yield 42;
-}
-
-async function* asyncGeneratorMethod() {
-  yield 42;
-}
-
-module.exports.generatorMethod = generatorMethod;
-module.exports.asyncGeneratorMethod = asyncGeneratorMethod;
+module.exports = {
+  RESPONSE: 'not-yet-received',
+};