+++ packages/jest-jasmine2/src/__tests__/itTestError.test.ts
@@ -14,11 +14,13 @@ describe('test/it error throwing', () => {
       'Missing second argument. It must be a callback function. Perhaps you want to use `test.todo` for a test placeholder.',
     );
   });
-  it("it throws an error when first argument isn't a string", () => {
+  it("it throws an error when first argument isn't valid", () => {
     expect(() => {
       // @ts-expect-error
       it(() => {});
-    }).toThrowError('Invalid first argument, () => {}. It must be a string.');
+    }).toThrowError(
+      'Invalid first argument, () => {}. It must be a named class, named function, number, or string.',
+    );
   });
   it('it throws an error when callback function is not a function', () => {
     expect(() => {
@@ -35,11 +37,13 @@ describe('test/it error throwing', () => {
       'Missing second argument. It must be a callback function. Perhaps you want to use `test.todo` for a test placeholder.',
     );
   });
-  test("test throws an error when first argument isn't a string", () => {
+  test("test throws an error when first argument isn't valid", () => {
     expect(() => {
       // @ts-expect-error
       test(() => {});
-    }).toThrowError('Invalid first argument, () => {}. It must be a string.');
+    }).toThrowError(
+      'Invalid first argument, () => {}. It must be a named class, named function, number, or string.',
+    );
   });
   test('test throws an error when callback function is not a function', () => {
     expect(() => {