+++ packages/jest-jasmine2/src/jasmine/Env.ts
@@ -31,7 +31,8 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 /* eslint-disable sort-keys, local/prefer-spread-eventually, local/prefer-rest-params-eventually */
 
 import {AssertionError} from 'assert';
-import {ErrorWithStack, isPromise} from 'jest-util';
+import type {Circus} from '@jest/types';
+import {ErrorWithStack, convertDescriptorToString, isPromise} from 'jest-util';
 import assertionErrorMessage from '../assertionErrorMessage';
 import isError from '../isError';
 import queueRunner, {
@@ -59,7 +60,11 @@ export default function jasmineEnv(j$: Jasmine) {
     fail: (error: Error | AssertionErrorWithStack) => void;
     pending: (message: string) => void;
     afterAll: (afterAllFunction: QueueableFn['fn'], timeout?: number) => void;
-    fit: (description: string, fn: QueueableFn['fn'], timeout?: number) => Spec;
+    fit: (
+      description: Circus.TestNameLike,
+      fn: QueueableFn['fn'],
+      timeout?: number,
+    ) => Spec;
     throwingExpectationFailures: () => boolean;
     randomizeTests: (value: unknown) => void;
     randomTests: () => boolean;
@@ -69,7 +74,7 @@ export default function jasmineEnv(j$: Jasmine) {
       suiteTree?: Suite,
     ) => Promise<void>;
     fdescribe: (
-      description: string,
+      description: Circus.TestNameLike,
       specDefinitions: SpecDefinitionsFn,
     ) => Suite;
     spyOn: (
@@ -84,18 +89,26 @@ export default function jasmineEnv(j$: Jasmine) {
     afterEach: (afterEachFunction: QueueableFn['fn'], timeout?: number) => void;
     clearReporters: () => void;
     addReporter: (reporterToAdd: Reporter) => void;
-    it: (description: string, fn: QueueableFn['fn'], timeout?: number) => Spec;
+    it: (
+      description: Circus.TestNameLike,
+      fn: QueueableFn['fn'],
+      timeout?: number,
+    ) => Spec;
     xdescribe: (
-      description: string,
+      description: Circus.TestNameLike,
       specDefinitions: SpecDefinitionsFn,
     ) => Suite;
-    xit: (description: string, fn: QueueableFn['fn'], timeout?: number) => Spec;
+    xit: (
+      description: Circus.TestNameLike,
+      fn: QueueableFn['fn'],
+      timeout?: number,
+    ) => Spec;
     beforeAll: (beforeAllFunction: QueueableFn['fn'], timeout?: number) => void;
     todo: () => Spec;
     provideFallbackReporter: (reporterToAdd: Reporter) => void;
     allowRespy: (allow: boolean) => void;
     describe: (
-      description: string,
+      description: Circus.TestNameLike,
       specDefinitions: SpecDefinitionsFn,
     ) => Suite;
 
@@ -364,7 +377,7 @@ export default function jasmineEnv(j$: Jasmine) {
         return spyRegistry.spyOn.apply(spyRegistry, args);
       };
 
-      const suiteFactory = function (description: string) {
+      const suiteFactory = function (description: Circus.TestNameLike) {
         const suite = new j$.Suite({
           id: getNextSuiteId(),
           description,
@@ -378,7 +391,10 @@ export default function jasmineEnv(j$: Jasmine) {
         return suite;
       };
 
-      this.describe = function (description: string, specDefinitions) {
+      this.describe = function (
+        description: Circus.TestNameLike,
+        specDefinitions,
+      ) {
         const suite = suiteFactory(description);
         if (specDefinitions === undefined) {
           throw new Error(
@@ -483,7 +499,7 @@ export default function jasmineEnv(j$: Jasmine) {
       }
 
       const specFactory = (
-        description: string,
+        description: Circus.TestNameLike,
         fn: QueueableFn['fn'],
         suite: Suite,
         timeout?: number,
@@ -534,11 +550,7 @@ export default function jasmineEnv(j$: Jasmine) {
       };
 
       this.it = function (description, fn, timeout) {
-        if (typeof description !== 'string') {
-          throw new Error(
-            `Invalid first argument, ${description}. It must be a string.`,
-          );
-        }
+        description = convertDescriptorToString(description);
         if (fn === undefined) {
           throw new Error(
             'Missing second argument. It must be a callback function. Perhaps you want to use `test.todo` for a test placeholder.',